Переключение на https, если указан только он + README + пример конфига + поддержка больше 255 провайдеров

+++ libs/config_epg/reader.go
@@ -29,8 +29,9 @@ func Load() {
     // Обновление списка каналов = 6 часов
     if p.UpdMins == 0 { p.UpdMins = 360 }
     // Порядок = 100 + порядок элемента
-    _default_order := 100 + uint8(i)
+    _default_order := 100 + uint16(i)
     if _default_order > 255 {  _default_order = 255 }
+
     if p.OrderEpg == 0 { p.OrderEpg = _default_order }
     if p.OrderIco == 0 { p.OrderIco = _default_order }
